requirements.txt
flask
flask-cors
pymongo
urllib3
PyYAML